+++ src/components/flowchart/style.css
@@ -1,6 +1,9 @@
-#canvas {
-    border: 1px solid #a3a3a3;
-    position: absolute;
+#svg {
+    background-image: linear-gradient(90deg, #dfdfdf 1px, transparent 0), linear-gradient(180deg, #dfdfdf 1px, transparent 0);
+    background-size: 20px 20px;
+    background-position: left -1px top -1px;
+    height: 100%;
+    width: 100%;
 }
 
 #chart {
@@ -11,7 +14,7 @@
     background-image: linear-gradient(to right, #f1f1f1 1px, transparent 1px),
     linear-gradient(to bottom, #f1f1f1 1px, transparent 1px);
     background-position: left -1px top -1px;
-    border: 1px solid #f1f1f1;
+    border: 1px solid #dfdfdf;
 }
 
 #position {
